Enums refactored
